+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # ???
 
+- Better support for Go, C++, C#, Haskell, Rust
+- Adds new "Sprinkles" theme
 - Switch line highlight BG to use text color instead of accent color since it
   made diff views hard to understand when the accent color is green
 - Tweaks and rearranges some colors in Periwinkle